requirements.txt
openai==0.15.0
python-dotenv==0.19.2
fastapi
uvicorn[standard]
sentence-transformers
torch
torch>=1.10
sentence-transformers>=2.2
